requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
openai
python-dotenv
gunicorn
captcha
